requirements.txt
bcrypt==4.2.1
dnspython==2.7.0
grpcio==1.68.1
grpcio-tools==1.68.1
protobuf==5.29.1
PyJWT==2.10.1
pymongo==4.10.1
setuptools==75.6.0
